# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8090
info:
  contact:
    name: PopuGate Support
    url: https://github.com/fussraider/PopuGate
  description: PopuGate is a Telegram MTProto proxy manager with a REST API.
  license:
    name: MIT
    url: https://github.com/fussraider/PopuGate/blob/main/LICENSE
  title: PopuGate API
  version: "1.0"
paths:
...
  /docker/install:
    post:
      consumes:
      - application/json
      description: Ensures Docker is installed on the host system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Install Docker
      tags:
      - docker
  /docker/status:
    get:
      consumes:
      - application/json
      description: Returns Docker installation and running status along with server
        version
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Docker status
      tags:
      - docker
  /engine/build:
    post:
      consumes:
      - application/json
      description: Builds the telemt engine Docker image from source. Supports optional
        force rebuild
      parameters:
      - description: Build options
        in: body
        name: body
        schema:
          properties:
            force:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Build engine image
      tags:
      - engine
...
  /engine/status:
    get:
      consumes:
      - application/json
      description: Returns the telemt engine image status, installed version, and
        whether it is up to date
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Engine status
      tags:
      - engine
...
  /geoblock:
    get:
      consumes:
      - application/json
      description: Returns the current geo-blocking mode and list of blocked/allowed
        countries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get geo-block settings
      tags:
      - geoblock
  /geoblock/add:
    post:
      consumes:
      - application/json
      description: Adds a country (ISO 3166-1 alpha-2 code) to the geo-block list
        and applies iptables rules
      parameters:
      - description: Country code to add
        in: body
        name: body
        required: true
        schema:
          properties:
            country:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add country to blocklist
      tags:
      - geoblock
  /geoblock/clear:
    post:
      consumes:
      - application/json
      description: Removes all countries from the geo-block list and clears iptables
        rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Clear all geo-block rules
      tags:
      - geoblock
  /geoblock/mode:
    put:
      consumes:
      - application/json
      description: Sets the geo-blocking mode to either blacklist or whitelist and
        re-applies rules
      parameters:
      - description: Mode to set (blacklist or whitelist)
        in: body
        name: body
        required: true
        schema:
          properties:
            mode:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set geo-block mode
      tags:
      - geoblock
  /geoblock/remove:
    post:
      consumes:
      - application/json
      description: Removes a country (ISO 3166-1 alpha-2 code) from the geo-block
        list and re-applies rules
      parameters:
      - description: Country code to remove
        in: body
        name: body
        required: true
        schema:
          properties:
            country:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove country from blocklist
      tags:
      - geoblock
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"